### 2018-02-21 #572 @Kevin-Mattheus-Moerman
@asmeurer @mattpap is this your cup of tea :tea: ? Would you be able to help review this [JOSS](http://joss.theoj.org/) submission? Alternatively, could you help suggest reviewers? Thanks! :rocket:

You can access the paper [here](https://github.com/openjournals/joss-papers/blob/joss.00572/joss.00572/10.21105.joss.00572.pdf).

The project is housed [here](https://github.com/HPAC/matchpy).

If you choose to accept this invitation we will initiate a review issue where you can tick boxes after reviewing the software (see also: http://joss.theoj.org/about#reviewer_guidelines).
