### 2018-02-23 c466098
Added some future work items to README.rst
