### 2018-03-05 #572 @henrik227
@Kevin-Mattheus-Moerman Can this be _anyone_ who is knowledgeable on the topic, or does the person has to be associated to JOSS in any way?
