### 2018-03-05 #572 @Kevin-Mattheus-Moerman
@henrik227 absolutely anybody with the required expertise, no affiliation with JOSS required.
