### 2018-03-19 #572 @henrik227
@Kevin-Mattheus-Moerman How do we proceed from here? I just checked a few authors of papers that were relevant for our work, but they don't seem to be on GitHub. The only other candidates I can think of right now are other SymPy developers, or somebody from the Mathics project (https://github.com/mathics), but I don't know any of them.
