### 2018-03-27 #572 @mrocklin
I am no longer active in this field and will have to pass.  I'll recommend
@llllllllll though.

On Tue, Mar 27, 2018 at 1:07 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @betatim <https://github.com/betatim> @asmeurer
> <https://github.com/asmeurer> @mattpap <https://github.com/mattpap> @cmry
> <https://github.com/cmry> @mrocklin <https://github.com/mrocklin> @rlamy
> <https://github.com/rlamy> @Suor <https://github.com/Suor> Would you be
> able to help review this JOSS <http://joss.theoj.org/> submission about
> pattern matching for symbolic computation? Alternatively, could you help
> suggest reviewers? Thanks! 🚀
>
> You can access the paper here
> <https://github.com/openjournals/joss-papers/blob/joss.00572/joss.00572/10.21105.joss.00572.pdf>
> .
>
> The project is housed here <https://github.com/HPAC/matchpy>.
>
> If you choose to accept this invitation we will initiate a review issue
> where you can tick boxes after reviewing the software (see also:
> http://joss.theoj.org/about#reviewer_guidelines).
>
