### 2018-03-28 #572 @Kevin-Mattheus-Moerman
@scopatz Please :rabbit: would you be able to help review this [JOSS](http://joss.theoj.org/) submission about pattern matching for symbolic computation? Alternatively, could you help suggest reviewers? Thanks! :rocket:

You can access the paper [here](https://github.com/openjournals/joss-papers/blob/joss.00572/joss.00572/10.21105.joss.00572.pdf).

The project is housed [here](https://github.com/HPAC/matchpy).

If you choose to accept this invitation we will initiate a review issue where you can tick boxes after reviewing the software (see also: http://joss.theoj.org/about#reviewer_guidelines).
