### 2018-03-29 #572 @Kevin-Mattheus-Moerman
@llllllllll we do not set a limit to the review time. We ask that reviewers work in a timely manor but we place no strict demands on this. There is quite a bit of variation in the time required to complete the review process. Typically a review process takes a couple of weeks. However I have also seen it completed in several days or several months (especially if significant changes are requested).
We'll hopefully be able to attract multiple reviewers (let me know if you know more suitable ones) which will start at different time points. So even if you cannot start immediately now you can certainly join in later. How much time do you think you'll need?
