### 2018-04-10 e6211ee
Added missing call to super().__init__()
