### 2018-04-11 #670 @whedon
[REVIEW]: MatchPy: Pattern Matching in Python
**Submitting author:** @henrik227 (<a href="http://orcid.org/0000-0001-6744-3605">Henrik Barthels</a>)
**Repository:** <a href="https://github.com/HPAC/matchpy" target ="_blank">https://github.com/HPAC/matchpy</a>
**Version:** v0.4.5
