### 2018-04-14 39dabdc
Fix doc build by removing typehint support
