### 2018-06-10 HPAC/matchpy#32 @llllllllll
clarify statement of need
Opening this issue as part of the JOSS paper review:

The statement of need in the documentation/README isn't very clear. For people unfamiliar with mathematica's pattern matching the overview doesn't really explain what this does. I would suggest putting something like the first paragraph of the summary section from the JOSS paper in the overview section of the README.

Also, if you just go the the docs page without first reading the README you would probably be a little lost. I would recommend making the README the index of documentation. You can make sure they stay in sync and still have the table of contents by using: `.. include:: ../../README.rst` in `index.rst` which will copy the readme into file.
