### 2018-06-10 #670 @llllllllll
The statement of need in the documentation could be more clear: https://github.com/HPAC/matchpy/issues/32. This should be a very small documentation change.

I only see a DOI for the sympy paper reference, I am not sure how these get assigned so this may be the only citation that has one. I am leaving this unchecked for now because of this.

After these two things are addressed this meets all of the criteria.
