### 2018-06-12 413ca24
Changed PyPI badge.
