### 2018-06-12 #670 @henrik227
I was about to look into this. Thanks for the comments, @llllllllll !
