### 2018-06-12 2546aad
Added simple example for pattern matching.
