### 2018-06-12 9ac3857
Using readme.rst as an introduction for the docs.
