+++ docs/index.rst
@@ -1,4 +1,4 @@
-.. include:: ../../README.rst
+.. include:: ../README.rst
 
 Table of Contents
 ==============
