### 2018-06-13 HPAC/matchpy#32 @henrik227
Thanks for the tip with the include. Does this look ok to you, @llllllllll ? https://matchpy.readthedocs.io/en/latest/
