### 2018-06-13 HPAC/matchpy#32 @llllllllll
Yeah, this looks great!
