### 2018-06-13 ba61429
Added missing DOIs.
