+++ joss_paper/paper.bib
@@ -60,7 +60,8 @@
     year      = 2017,
     month     = nov,
     note      = "In conjunction with SC17: The International Conference for High Performance Computing, Networking, Storage and Analysis",
-    url       = "https://arxiv.org/pdf/1710.00077.pdf"
+    url       = "https://arxiv.org/pdf/1710.00077.pdf",
+    doi       = {10.1145/3149869.3149871}
 }
 
 @inproceedings{krebber2017:3,
@@ -69,5 +70,6 @@
     booktitle = "Proceedings of the 15th Python in Science Conference",
     year      = 2017,
     month     = jul,
-    url       = "http://conference.scipy.org/proceedings/scipy2017/pdfs/manuel_krebber.pdf"
+    url       = "http://conference.scipy.org/proceedings/scipy2017/pdfs/manuel_krebber.pdf",
+    doi       = {10.25080/shinma-7f4c6e7-00b}
 }
