### 2018-06-13 #670 @henrik227
I improved the statement of need and added DOIs for all references that have one (at least to my knowledge).
