### 2018-06-14 a073b94
Added JOSS badge.
