### 2018-06-15 #670 @henrik227
What happens next, now that the checklist is complete?
