### 2018-06-15 #670 @Kevin-Mattheus-Moerman
@henrik227, once @llllllllll confirms, could you create a DOI for an archived version of the reviewed software, and list it here? You can use a service like [Zenodo](http://zenodo.org/) to do this (see also [here](https://guides.github.com/activities/citable-code/)).
