### 2018-06-15 #670 @llllllllll
I confirm that this meets all of criteria for a submission.

I would also like to say that this is a very interesting and useful project. I'm glad I was able to learn more about this while reviewing.
