### 2018-06-20 #670 @henrik227
Is it ok to create a new MatchPy release (0.4.5), so I can make use of Zenodo automatically archiving releases? And if yes, do I also change that in the metadata of the paper?
