### 2018-06-20 #670 @Kevin-Mattheus-Moerman
Yes you can, so go ahead and create the new release and list the DOI for it here. @arfon can then help update the version number listed here and the metadata for the paper.
