### 2018-06-21 #670 @henrik227
The DOI of Version v0.4.5 is "10.5281/zenodo.1294930"
