### 2018-06-21 1dda07b
Open README with correct encoding.
