+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup, find_packages
 
 root = os.path.dirname(__file__)
 
-with open(os.path.join(root, 'README.rst')) as f:
+with open(os.path.join(root, 'README.rst'), encoding='utf-8') as f:
     readme = f.read()
 
 
