### 2018-06-21 #670 @Kevin-Mattheus-Moerman
@arfon can you set the version to 0.4.5, assign [10.5281/zenodo.1294930](https://zenodo.org/record/1294930) as the DOI, and proceed to accept this submission? Thanks :rocket: :rabbit: :camel: :tulip:
